rust: amortize allocations
This is a PoC for only the Rust code that shows how to optimize it a bit
more by both amortizing allocation and reusing the result of UTF-8
decoding instead of repeating it for each iteration of the inner loop.

The results on my machine:

    $ yarn bench
    yarn run v1.22.4
    $ node run.js bench
    go: 1.573124
    javascript: 2.94
    rust: 1.295147587
    Done in 6.17s.

Of course, this isn't a fair comparison, since this doesn't update the
Go code to reuse allocation either. Although it wouldn't be surprising
if its GC was doing this automatically anyway.

I'm not necessarily submitting this with a request to merge it, but
rather, just providing more data.